% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFE.R
\name{autoFE}
\alias{autoFE}
\title{autoFE}
\usage{
autoFE(train, valid, test, x, y, savePath, verbose = FALSE)
}
\arguments{
\item{train}{data.table object}

\item{valid}{data.table object}

\item{test}{data.table object}

\item{savePath}{save path}
}
\description{
autoFE
}
\examples{
library(data.table)
library(rAutoFE)
library(rAutoFS)
library(h2o)
savePath <- "c:/tmp"
dir.create(savePath)
data(churn)
churn <- as.data.table(churn)
churn[, Area.Code:=as.factor(Area.Code)]
splits <- splitFrame(dt=churn, ratio = c(0.5, 0.2), seed=1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
y = "Churn."
x = colnames(train)[colnames(train)!=y]
h2o.init()
dataset <- autoFE(train=train, valid=valid, test=test, x=x, y=y, savePath=savePath, verbose=TRUE)
}
