% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestNormalize2.R
\name{bestNormalize_transform}
\alias{bestNormalize_transform}
\title{bestNormalize_transform}
\usage{
bestNormalize_transform(dt, fit)
}
\arguments{
\item{dt}{data.table object}
}
\description{
bestNormalize_transform
}
\examples{
library(bestNormalize)
library(data.table)
library(rAutoFE)
library(rAutoFS)
library(h2o)
churn <- as.data.table(churn)
splits <- rAutoFE::splitFrame(dt=churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
h2o.init()
data_hex <- as.h2o(rbind(train, valid))
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
ml <- autoFS(data_hex, x, y, num_of_model=5, num_of_vi=10)
fit <- bestNormalize_fit(dt=train, vi=ml$top_vi)
train <- bestNormalize_transform(dt=train, fit=fit)
valid <- bestNormalize_transform(dt=valid, fit=fit)
test  <- bestNormalize_transform(dt=test, fit=fit)
}
