% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin4numeric.R
\name{bin4numeric_transform}
\alias{bin4numeric_transform}
\title{bin4numeric_transform}
\usage{
bin4numeric_transform(dt, fit)
}
\arguments{
\item{dt}{data.table object}
}
\description{
bin4numeric_transform
}
\examples{
library(data.table)
library(binr)
library(rAutoFE)
library(rAutoFS)
library(h2o)
data(churn)
churn <- as.data.table(churn)
splits <- rAutoFE::splitFrame(dt=churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
h2o.init()
data_hex <- as.h2o(rbind(train, valid))
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
ml <- autoFS(data_hex, x, y, num_of_model=5, num_of_vi=10)
fit <- bin4numeric_fit(dt=train, vi=ml$top_vi)
bin4numeric_transform(dt=train, fit=fit)
bin4numeric_transform(dt=valid, fit=fit)
bin4numeric_transform(dt=test, fit=fit)
}
