% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{convert2fac}
\alias{convert2fac}
\title{convert2fac}
\usage{
convert2fac(dt, columns)
}
\arguments{
\item{dt}{data.table object}

\item{columns}{character vector}
}
\description{
convert all selected features to factor
}
