% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvNumtoCat_WoE.R
\name{cvNumtoCat_WoE_transform}
\alias{cvNumtoCat_WoE_transform}
\title{cvNumtoCat_WoE_transform}
\usage{
cvNumtoCat_WoE_transform(data, fit)
}
\arguments{
\item{data}{data.table object}
}
\description{
Convert the numeric feature to categorical feature by cross validated mean criteria then make WeO feature
}
\examples{
library(data.table)
library(binr)
library(rAutoFE)
library(rAutoFS)
library(h2o)
library(caret)
data(churn)
churn <- as.data.table(churn)
splits <- rAutoFE::splitFrame(dt=churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
h2o.init()
data_hex <- as.h2o(rbind(train, valid))
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
ml <- autoFS(data_hex, x, y, num_of_model=5, num_of_vi=20)
vi <- ml$top_vi
fit <- cvNumtoCat_WoE_fit(data=train, vi=vi, y, bin = 10, k = 5)
cvNumtoCat_WoE_transform(data=train, fit=fit)
cvNumtoCat_WoE_transform(data=valid, fit=fit)
cvNumtoCat_WoE_transform(data=test, fit=fit)
}
