% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{dt_convert2fac}
\alias{dt_convert2fac}
\title{dt_convert2fac}
\usage{
dt_convert2fac(dt, columns)
}
\arguments{
\item{dt}{data.table object}

\item{columns}{character vector}
}
\description{
convert all selected features to factor
}
