% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{dt_splitFrame}
\alias{dt_splitFrame}
\title{dt_splitFrame}
\usage{
dt_splitFrame(dt, ratio, seed)
}
\arguments{
\item{dt}{data.table object}

\item{ratio}{numeric vector}

\item{seed}{integer}
}
\description{
data split using data.table
}
