% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorAnalysisEncoding.R
\name{errorAnalysisEncoding_transform}
\alias{errorAnalysisEncoding_transform}
\title{errorAnalysisEncoding_transform}
\usage{
errorAnalysisEncoding_transform(data, fit)
}
\arguments{
\item{data}{data.table object. It must contains 'error' columns which has binary variables(0,1).}

\item{fit}{list of information for fitting}
}
\description{
transform for error Analysis.
}
\examples{
library(rAutoFE)
data(churn)
churn <- as.data.table(churn)
churn[,'error'] <- ifelse(churn[,Churn.] == "False.", 0, 1)
churn$Account.Length[1:500] <- NA
splits <- rAutoFE::splitFrame(dt = churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test <- splits[[3]]
fit_info <- errorAnalysisEncoding_fit(data = train, column_name = c("Day.Charge", "Account.Length", "State"), fit_type = c("size", "width", NA), x_type = c("num", "num", "cat"), bins = c(5, 6, NA))
train <- errorAnalysisEncoding_transform(data = train, fit = fit_info)
valid <- errorAnalysisEncoding_transform(data = valid, fit = fit_info)
test <- errorAnalysisEncoding_transform(data = test, fit = fit_info)
}
