% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCombination.R
\name{featureComb_fit_transform}
\alias{featureComb_fit_transform}
\title{featureComb_fit_transform}
\usage{
featureComb_fit_transform(dt, vi)
}
\arguments{
\item{dt}{data.table object}
}
\description{
featureComb_fit_transform
}
\examples{
library(data.table)
library(binr)
library(rAutoFE)
library(rAutoFS)
library(h2o)
churn <- as.data.table(churn)
splits <- rAutoFE::splitFrame(dt=churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
h2o.init()
data_hex <- as.h2o(rbind(train, valid))
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
ml <- autoFS(data_hex, x, y, num_of_model=5, num_of_vi=10)
train <- featureComb_fit_transform(dt=train, vi=ml$top_vi)
valid <- featureComb_fit_transform(dt=valid, vi=ml$top_vi)
test  <- featureComb_fit_transform(dt=test, vi=ml$top_vi)
}
