% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyEncoding.R
\name{frequencyEncoding_fit}
\alias{frequencyEncoding_fit}
\title{frequencyEncoding_fit}
\usage{
frequencyEncoding_fit(data, x)
}
\examples{
source("INTERNAL/auto_fe/R/user_functions.R")
library(data.table)
path = c("./INTERNAL/auto_fe/data")
data <- readRDS(file.path(path,"churn.Rda"))
data <- dt_convert2fac(data=data, columns=c("Churn.","State","Area.Code","Int.l.Plan","VMail.Plan"))
splits <- dt_splitFrame(data=data, ratio = c(0.5, 0.2), seed=1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- frequencyEncoding_fit(data = train, x = x)
saveRDS(fit,"./bak/fit.Rda")
}
