% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyEncoding.R
\name{frequencyEncoding_transform}
\alias{frequencyEncoding_transform}
\title{frequencyEncoding_transform}
\usage{
frequencyEncoding_transform(data, x = x, fit = fit)
}
\examples{
fit <- readRDS("./bak/fit.Rda")
train <- frequencyEncoding_transform(data = train, x = x, fit = fit)
valid <- frequencyEncoding_transform(data = valid, x = x, fit = fit)
test  <- frequencyEncoding_transform(data = test, x = x, fit = fit)
}
