% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{removeOneLevel_transform}
\alias{removeOneLevel_transform}
\title{removeOneLevel_transform}
\usage{
removeOneLevel_transform(dt, fit)
}
\arguments{
\item{dt}{data.table object}
}
\description{
removeOneLevel_transform
}
