% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoding.R
\name{targetEncoding_fit}
\alias{targetEncoding_fit}
\title{targetEncoding_fit}
\usage{
targetEncoding_fit(dt, x, y)
}
\examples{
library(rAutoFE)
library(data.table)
data(churn, package = "rAutoFE")
str(churn)
data.table::setDT(churn)
splits <- dt_splitFrame(data=churn, ratio = c(0.5, 0.2), seed=1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- frequencyEncoding_fit(dt = train, x = x)
}
