% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoding.R
\name{targetEncoding_transform}
\alias{targetEncoding_transform}
\title{targetEncoding_transform}
\usage{
targetEncoding_transform(dt, x, y, fit)
}
\examples{
library(rAutoFE)
library(data.table)
data(churn, package = "rAutoFE")
str(churn)
data.table::setDT(churn)
splits <- dt_splitFrame(data=churn, ratio = c(0.5, 0.2), seed=1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- frequencyEncoding_fit(dt = train, x = x)
saveRDS(fit,"fit.rds")
rm(fit)
fit <- readRDS("fit.rds")
train <- frequencyEncoding_transform(dt = train, x = x, fit = fit)
valid <- frequencyEncoding_transform(dt = valid, x = x, fit = fit)
test  <- frequencyEncoding_transform(dt = test, x = x, fit = fit)
}
