% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightOfEvidence.R
\name{WoE_fit}
\alias{WoE_fit}
\title{WoE_fit}
\usage{
WoE_fit(data, target.var, pred.var = NA)
}
\arguments{
\item{data}{data.table object.}

\item{target.var}{A character representing the name of the binary variable}

\item{pred.var}{character vector column names}
}
\description{
train for WoE and information value.
}
\examples{
library(rAutoFE)
data(churn)
churn <- as.data.table(churn)
churn[,'target'] <- ifelse(churn[,Churn.] == "False.", 0, 1)
churn$Account.Length[1:500] <- NA
splits <- splitFrame(dt = churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test <- splits[[3]]
fit_info <- WoE_fit(data = train, target.var = "target", pred.var = c("Day.Charge", "Account.Length", "State"))
}
