% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightOfEvidence.R
\name{WoE_transform}
\alias{WoE_transform}
\title{WoE_transform}
\usage{
WoE_transform(data, fit)
}
\arguments{
\item{data}{data.table object.}

\item{fit}{list of information for fitting}
}
\description{
transform for WoE and information value.
}
\examples{
library(rAutoFE)
data(churn)
churn <- as.data.table(churn)
churn[,'error'] <- ifelse(churn[,Churn.] == "False.", 0, 1)
churn$Account.Length[1:500] <- NA
splits <- rAutoFE::splitFrame(dt = churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test <- splits[[3]]
fit_info <- WoE_fit(data = train, target.var = "target", pred.var = c("Day.Charge", "Account.Length", "State"))
train <- WoE_transform(data = train, fit = fit_info)
valid <- WoE_transform(data = valid, fit = fit_info)
test <- WoE_transform(data = test, fit = fit_info)
}
