% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorAnalysisEncoding.R
\name{errorAnalysisEncoding_fit}
\alias{errorAnalysisEncoding_fit}
\title{errorAnalysisEncoding_fit}
\usage{
errorAnalysisEncoding_fit(data, ml, y, bins = 30, numVar = 10)
}
\arguments{
\item{data}{data.table object. It must contains 'error' columns which has binary variables(0,1).}
}
\description{
train for error Analysis.
}
