% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorAnalysisEncoding.R
\name{errorAnalysisEncoding_transform}
\alias{errorAnalysisEncoding_transform}
\title{errorAnalysisEncoding_transform}
\usage{
errorAnalysisEncoding_transform(data, fit)
}
\arguments{
\item{data}{data.table object. It must contains 'error' columns which has binary variables(0,1).}
}
\description{
transform for error Analysis.
}
