% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyEncoding.R
\name{frequencyEncoding_fit}
\alias{frequencyEncoding_fit}
\title{frequencyEncoding_fit}
\usage{
frequencyEncoding_fit(dt, x)
}
\arguments{
\item{dt}{data.table object}

\item{x}{character vector of independet variables}
}
\description{
train frequen-based encoding for categorical features
}
\examples{
library(rAutoFE)
library(data.table)
data(churn, package = "rAutoFE")
data.table::setDT(churn)
splits <- rAutoFE::splitFrame(dt=churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- rAutoFE::frequencyEncoding_fit(dt = train, x = x)
}
