% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoding.R
\name{targetEncoding_fit}
\alias{targetEncoding_fit}
\title{targetEncoding_fit}
\usage{
targetEncoding_fit(dt, x, y)
}
\arguments{
\item{dt}{data.table object}

\item{x}{character vector of independet variables}

\item{y}{character string of dependent variable}
}
\description{
train for target feature encoding
}
\examples{
library(rAutoFE)
library(data.table)
data(churn, package = "rAutoFE")
data.table::setDT(churn)
splits <- splitFrame(dt = churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- rAutoFE::targetEncoding_fit(dt = train, x = x, y = y)
}
