% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoding.R
\name{targetEncoding_fit}
\alias{targetEncoding_fit}
\title{targetEncoding_fit}
\usage{
targetEncoding_fit(data, x, y)
}
\examples{
require(data.table)
require(h2o)
h2o.init()
data <- fread("data/churn.csv")
data <- dt_convert2fac(data=data, columns=c("Churn.","State","Area.Code","Int.l.Plan","VMail.Plan"))
splits <- dt_splitFrame(data=data, ratio = c(0.5, 0.2), seed=1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- targetEncoding_fit(data = train, x = x, y = y)
saveRDS(fit,"fit.Rda")
}
