% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoding.R
\name{targetEncoding_transform}
\alias{targetEncoding_transform}
\title{targetEncoding_transform}
\usage{
targetEncoding_transform(dt, fit)
}
\arguments{
\item{dt}{data.table object}

\item{fit}{targetEncoding_fit object}

\item{x}{character vector of independet variables}

\item{y}{character string of dependent variable}
}
\description{
transform dataset using fit
}
\examples{
library(rAutoFE)
library(data.table)
data(churn, package = "rAutoFE")
data.table::setDT(churn)
splits <- rAutoFE::splitFrame(dt = churn, ratio = c(0.5, 0.3), seed = 1234)
train <- splits[[1]]
valid <- splits[[2]]
test  <- splits[[3]]
x = c("State","Area.Code")
y = "Churn."
fit <- rAutoFE::targetEncoding_fit(dt = train, x = x, y = y)
saveRDS(fit,"fit.rds")
rm(fit)
fit <- readRDS("fit.rds")
train <- rAutoFE::targetEncoding_transform(dt = train, fit = fit)
valid <- rAutoFE::targetEncoding_transform(dt = valid, fit = fit)
test  <- rAutoFE::targetEncoding_transform(dt = test, fit = fit)
}
