% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoding.R
\name{targetEncoding_transform}
\alias{targetEncoding_transform}
\title{targetEncoding_transform}
\usage{
targetEncoding_transform(data, x, y, fit)
}
\examples{
fit <- readRDS("./bak/fit.Rda")
train <- targetEncoding_transform(data = train, x = x, y = y, fit = fit)
valid <- targetEncoding_transform(data = valid, x = x, y = y, fit = fit)
test  <- targetEncoding_transform(data = test, x = x, y = y, fit = fit)
}
